src/main/webapp/app/entities/sens/route/sens-routing-resolve.service.spec.ts
jest.mock('@angular/router');

import { TestBed } from '@angular/core/testing';
import { HttpResponse } from '@angular/common/http';
import { HttpClientTestingModule } from '@angular/common/http/testing';
import { ActivatedRouteSnapshot, Router } from '@angular/router';
import { of } from 'rxjs';

import { ISens, Sens } from '../sens.model';
import { SensService } from '../service/sens.service';

import { SensRoutingResolveService } from './sens-routing-resolve.service';

describe('Sens routing resolve service', () => {
  let mockRouter: Router;
  let mockActivatedRouteSnapshot: ActivatedRouteSnapshot;
  let routingResolveService: SensRoutingResolveService;
  let service: SensService;
  let resultSens: ISens | undefined;

  beforeEach(() => {
    TestBed.configureTestingModule({
      imports: [HttpClientTestingModule],
      providers: [Router, ActivatedRouteSnapshot],
    });
    mockRouter = TestBed.inject(Router);
    mockActivatedRouteSnapshot = TestBed.inject(ActivatedRouteSnapshot);
    routingResolveService = TestBed.inject(SensRoutingResolveService);
    service = TestBed.inject(SensService);
    resultSens = undefined;
  });

  describe('resolve', () => {
    it('should return ISens returned by find', () => {
      // GIVEN
      service.find = jest.fn(id => of(new HttpResponse({ body: { id } })));
      mockActivatedRouteSnapshot.params = { id: 123 };

      // WHEN
      routingResolveService.resolve(mockActivatedRouteSnapshot).subscribe(result => {
        resultSens = result;
      });

      // THEN
      expect(service.find).toBeCalledWith(123);
      expect(resultSens).toEqual({ id: 123 });
    });

    it('should return new ISens if id is not provided', () => {
      // GIVEN
      service.find = jest.fn();
      mockActivatedRouteSnapshot.params = {};

      // WHEN
      routingResolveService.resolve(mockActivatedRouteSnapshot).subscribe(result => {
        resultSens = result;
      });

      // THEN
      expect(service.find).not.toBeCalled();
      expect(resultSens).toEqual(new Sens());
    });

    it('should route to 404 page if data not found in server', () => {
      // GIVEN
      jest.spyOn(service, 'find').mockReturnValue(of(new HttpResponse({ body: null as unknown as Sens })));
      mockActivatedRouteSnapshot.params = { id: 123 };

      // WHEN
      routingResolveService.resolve(mockActivatedRouteSnapshot).subscribe(result => {
        resultSens = result;
      });

      // THEN
      expect(service.find).toBeCalledWith(123);
      expect(resultSens).toEqual(undefined);
      expect(mockRouter.navigate).toHaveBeenCalledWith(['404']);
    });
  });
});
